requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytz